\name{rastafilt}
\alias{rastafilt}
\title{Using RASTA filter}
\description{
  \code{rastafilt} ist used to filter each frequency band with the RASTA (Relative Spectral Transform) filter.}
\usage{
  rastafilt(x)
}
\arguments{
  \item{x}{a matrix where rows of \code{x} are the critical bands and columns of \code{x} are the frames.}
}
\details{
}
\value{
  Matrix of the same type as \code{x} after RASTA filtering.
}
\references{
}
\author{
}
\note{
}

\seealso{{\code{\link{rastaplp}}}
\examples{
}
\concept{ filter }
\concept{ frequency }
\concept{ RASTA }
